Fixing build for Qt6 and for msvc in 64 bit

+++ lib/src/widgets/AbstractItemListWidget.cpp
@@ -146,7 +146,7 @@ int AbstractItemListWidget::addItem(
   emit itemCountChanged();
   emit currentIndexChanged();
 
-  const auto [bgColor, fgColor, badgeBgColor, badgeFgColor] = getItemBgAndFgColor(_items.size(), MouseState::Normal);
+  const auto [bgColor, fgColor, badgeBgColor, badgeFgColor] = getItemBgAndFgColor(static_cast<int>(_items.size()), MouseState::Normal);
   bgColorAnimation->setDuration(animDuration);
   bgColorAnimation->setStartValue(QVariant::fromValue<QColor>(bgColor));
   bgColorAnimation->setEndValue(QVariant::fromValue<QColor>(bgColor));
@@ -167,7 +167,7 @@ int AbstractItemListWidget::addItem(
   badgeFgColorAnimation->setDuration(animDuration);
   badgeFgColorAnimation->setEasingCurve(QEasingCurve::Type::InOutCubic);
 
-  return _items.size() - 1;
+  return static_cast<int>(_items.size()) - 1;
 }
 
 void AbstractItemListWidget::removeItem(int index) {
@@ -430,7 +430,7 @@ QSize AbstractItemListWidget::sizeHint() const {
     neededH = std::max(neededH, item.sizeHint.height());
   }
   const auto itemCount = _items.size();
-  const auto spacings = itemCount > 1 ? (itemCount - 1) * spacing : 0;
+  const auto spacings = static_cast<int>(itemCount > 1 ? (itemCount - 1) * spacing : 0);
   neededW += spacings + padding.left() + padding.right();
   neededH += padding.top() + padding.bottom();
